+++ NEWS.md
@@ -1,5 +1,10 @@
 # testthat (development version)
 
+* If you define a function called `testthat_state()`, testthat will call it
+  before and after every test and report if there are any differences. This
+  is very useful for detecting if any of your tests have made changes to 
+  global state (like options, env vars, or connections) (#1674).
+
 * Only report test files that take longer than a second (#1806).
 
 * Links generate for snapshot hints now work when using parallel tests (#1802).